Back End Updates
The ability to pull an image from the database utilising a post request from front end is now functional. It then converts it back into an image to have the ffmpeg commands be done to it. This is where stuff goes wrong. It must simply be a syntax error judging by the error that comes from ffmpeg. If this is fixed then the API should be fully functional, I just need to convert the resulting image back into a blob.