Back end transcoding now functional, kinda
These changes allow for an image to be created. Now, the problem is that the image is not supported to be opened so I don't know if it fully worked. The file is 40kb going down from 720p to 480p, so I think it worked somewhat, because that checks out.